backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-multipart==0.0.9
pydantic-settings==2.5.0
pydantic==2.9.0
boto3==1.35.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0
pytesseract==0.3.13
Pillow==10.4.0
